requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
playwright==1.43.0
pytest-playwright==0.4.4
